Crate rl_sys

This library provides native bindings for the GNU readline library.

The GNU Readline library provides a set of functions for use by applications that allow users to edit command lines as they are typed in. Both Emacs and vi editing modes are available. The Readline library includes additional functions to maintain a list of previously-entered command lines, to recall and perhaps reedit those lines, and perform csh-like history expansion on previous commands.

Structs

ReadlineError
Represents an error that has occurred within the API.

Functions

add_history
Wraps the libreadline add_history functionality. The argument is the line to add to history.

add_history_persist
Add the given line to readline history and persistently to a file at the given path. This is useful in conjunction with the preload_history function for keeping a useful history for your application.

preload_history
Preload the readline history with lines from the given file. This is often use in conjunction with the add_history_persist api to maintain a readline history persistently.

readline
Wraps the libreadline readline function. The argument is the prompt to use.

version
Generate a version string.