Function rl_sys::add_history_persist

pub fn add_history_persist(line: &str, file: &Path) -> Result<(), ReadlineError>

Add the given line to readline history and persistently to a file at the given path. This is useful in conjunction with the preload_history function for keeping a useful history for your application.

Note that this function will only add the line to the readline history and the file history if it doesn't already exist there.

Examples

use rl_sys;
use std::path::Path;

let history_file = Path::new("/home/user/.app-hist");
match rl_sys::add_history_persist("ls -al", &history_file) {
    Ok(_)  => println!("Success!"),
    Err(e) => println!("{}", e),
}