Module rl_sys::history::expand

2.3.7 History Expansion

These functions implement history expansion.

Functions

arg_extract
Extract a string segment consisting of the first through last arguments present in string s. Arguments are split using history_tokenize. If either first or last is < 0, then make that arg count from the right (subtract from the number of tokens, so that first = -1 means the next to last token on the line). If first and last are 36 (ASCII '$') the last arg from string s is used.

expand
Expand string, placing the result into output, a pointer to a string (see section 1.1 History Expansion).

get_event
Returns the text of the history event beginning at s + *idx. *idx is modified to point to after the event specifier. At function entry, *idx points to the index into string where the history event specification begins. add_delim is a character that is allowed to end the event specification in addition to the "normal" terminating characters.

tokenize
Return an array of tokens parsed out of string s, much as the shell might. The tokens are split on the characters in the history_word_delimiters variable, and shell quoting conventions are obeyed.