Function rl_sys::history::expand::arg_extract

pub fn arg_extract(s: &str, first: i32, last: i32) -> Result<String, HistoryError>

Extract a string segment consisting of the first through last arguments present in string s. Arguments are split using history_tokenize. If either first or last is < 0, then make that arg count from the right (subtract from the number of tokens, so that first = -1 means the next to last token on the line). If first and last are 36 (ASCII '$') the last arg from string s is used.

Examples

use rl_sys::history::expand;

let mut res = expand::arg_extract("one two three 'a b c' <def>", 0, 1).unwrap();
assert!(res == "one two");
// 36 is used to represent ASCII '$'.
res = expand::arg_extract("one two three", 36, 36).unwrap();
assert!(res == "three");