Function rl_sys::history::expand::expand

pub fn expand(s: &str) -> Result<(isize, String), HistoryError>

Expand string, placing the result into output, a pointer to a string (see section 1.1 History Expansion).

Returns:

0 - If no expansions took place (or, if the only change in the text was the removal of escape characters preceding the history expansion character).

1 - If expansions did take place.

-1 - If there was an error in expansion.

2 - If the returned line should be displayed, but not executed, as with the :p modifier (see section 1.1.3 Modifiers).

If an error occurred in expansion, then output contains a descriptive error message.

Examples

use rl_sys::history::{listmgmt, expand};

assert!(listmgmt::add("ls -al").is_ok());
let (res, out) = expand::expand("!ls").unwrap();
assert!(res == 1);
assert!(out == "ls -al");