Function rl_sys::history::expand::tokenize

pub fn tokenize(s: &str) -> Result<Vec<String>, HistoryError>

Return an array of tokens parsed out of string s, much as the shell might. The tokens are split on the characters in the history_word_delimiters variable, and shell quoting conventions are obeyed.

Examples

use rl_sys::history::expand;

let res = expand::tokenize("one two three 'a b c' <def>").unwrap();
// ["one", "two", "three", "'a b c'", "<", "def", ">"]
assert!(res.len() == 7);
assert!(res[0] == "one");
assert!(res[6] == ">");