Module rl_sys::history::listinfo

2.3.3 Information About the History List

These functions return information about the entire history list or individual list entries.

Functions

current
Return the history entry at the current position, as determined by where_history()`. If there is no entry there, return a HistoryError.

get
Return the history entry at position offset, starting from history_base. If there is no entry there, or if offset is greater than the history length, return a HistoryError.

get_time
Return the time stamp associated with the history entry entry.

list
Return a Vec which is the current input history. Element 0 of this list is the beginning of time. If there is no history, return an empty vector.

offset
Returns the offset of the current history element.

total_bytes
Return the number of bytes that the primary history entries are using. This function returns the sum of the lengths of all the lines in the history.