Function rl_sys::history::listinfo::get

pub fn get<'a>(offset: i32) -> Result<&'a mut HistoryEntry, HistoryError>

Return the history entry at position offset, starting from history_base. If there is no entry there, or if offset is greater than the history length, return a HistoryError.

Examples

use rl_sys::history::{listinfo, listmgmt, vars};

assert!(listmgmt::add("test").is_ok());
assert!(vars::history_base == 1);
assert!(listinfo::get(1).is_ok());