Function rl_sys::history::listmgmt::replace_entry

pub fn replace_entry(offset: i32, line: &str, appdata: Option<*mut c_void>) -> Result<&mut HistoryEntry, HistoryError>

Replace the history entry at offset with the given line and data. This returns the old entry so the caller can dispose of any application-specific data. In the case of an invalid offset, an Err is returned.

Examples

use rl_sys::history::{listmgmt, vars};

assert!(listmgmt::add("ls -al").is_ok());
assert_eq!(vars::history_length, 1);
assert!(listmgmt::replace_entry(0, "test", None).is_ok());
assert_eq!(vars::history_length, 1);