Module rl_sys::history::mgmt

2.3.1 Initializing History and State Management

This section describes functions used to initialize and manage the state of the History library when you want to use the history functions in your program.

Functions

cleanup
Cleanup the history state. This should be called on program exit or after you are completely finished using the History API.

get_state
Return a structure describing the current state of the input history.

init
Begin a session in which the history functions might be used. This initializes the interactive variables. Note that this will only call using_history once after first use.

set_state
Set the state of the history list according to state.