Module rl_sys::history::move_

2.3.4 Moving Around the History List

These functions allow the current index into the history list to be set or changed.

Functions

next
Move the current history offset forward to the next history entry, and return a pointer to that entry. If there is no next entry, return a HistoryError.

previous
Back up the current history offset to the previous history entry, and return a pointer to that entry. If there is no previous entry, return a HistoryError.

set_pos
Set the current history offset, an absolute index into the list. Returns true on success, false if the offset is less than zero or greater than the number of history entries.