Module rl_sys::history::search

2.3.5 Searching the History List

These functions allow searching of the history list for entries containing a specific string. Searching may be performed both forward and backward from the current history position. The search may be anchored, meaning that the string must match at the beginning of the history entry.

Enums

Direction
The direction to search through the history entries.

Functions

search
Search the history for string, starting at the current history offset. If direction is Backward, then the search is through previous entries, otherwise through subsequent entries. If string is found, then the current history index is set to that history entry, and the value returned is the offset in the line of the entry where string was found. Otherwise, nothing is changed, and a -1 is returned.

search_pos
Search for string in the history list, starting at pos, an absolute index into the list. If direction Backward, the search proceeds backward from pos, otherwise forward. Returns the absolute index of the history element where string was found, or -1 otherwise.

search_prefix
Search the history for string, starting at the current history offset. The search is anchored: matching lines must begin with string. If direction is Backward, then the search is through previous entries, otherwise through subsequent entries. If string is found, then the current history index is set to that entry, and the return value is 0. Otherwise, nothing is changed, and a -1 is returned.