Function rl_sys::history::search::search

pub fn search<T>(s: &str, dir: T) -> Result<isize, HistoryError> where T: Into<i32>

Search the history for string, starting at the current history offset. If direction is Backward, then the search is through previous entries, otherwise through subsequent entries. If string is found, then the current history index is set to that history entry, and the value returned is the offset in the line of the entry where string was found. Otherwise, nothing is changed, and a -1 is returned.

Examples

use rl_sys::history::listmgmt;
use rl_sys::history::search::{self, Direction};

assert!(listmgmt::add("ls -al").is_ok());
let res = search::search("-al", Direction::Forward).unwrap();
assert!(res == 3);
let res1 = search::search("blah", Direction::Forward).unwrap();
assert!(res1 < 0);