Function rl_sys::history::vars::get_search_delimiter_chars

pub fn get_search_delimiter_chars() -> Result<Vec<char>, HistoryError>

Get the list of search delimiter characters. The list of additional characters which can delimit a history search string, in addition to space, TAB, ':' and '?' in the case of a substring search. The default is empty.

Examples

use rl_sys::history::vars;

let delims: String = match vars::get_search_delimiter_chars() {
    Ok(v) => {
        v.into_iter().collect()
    },
    Err(_) => {
        String::new()
    },
};
assert!(delims.is_empty());