Function rl_sys::history::vars::set_inhibit_expansion_function

pub fn set_inhibit_expansion_function(f: InhibitExpansionFunc)

This should be set to a function of the following type: extern fn(*mut c_char, c_uint) -> c_int. The first argument is a string pointer and the second is an int index into that string. It should return a non-zero value if the history expansion starting at string[i] should not be performed; zero if the expansion should be done. It is intended for use by applications like Bash that use the history expansion character for additional purposes. By default, this variable is set to NULL.

Examples

use libc::{c_char, c_int, c_uint};
use rl_sys::history::vars;

extern fn blah(_: *mut c_char, _: c_uint) -> c_int {
    0
}

vars::set_inhibit_expansion_function(Some(blah));