Module rl_sys::readline::binding

2.4.3 Binding Keys

Key sequences are associated with functions through the keymap. Readline has several internal keymaps: emacs_standard_keymap, emacs_meta_keymap, emacs_ctlx_keymap, vi_movement_keymap, and vi_insertion_keymap. emacs_standard_keymap is the default, and the examples in this manual assume that.

Since readline() installs a set of default key bindings the first time it is called, there is always the danger that a custom binding installed before the first call to readline() will be overridden. An alternate mechanism is to install custom key bindings in an initialization function assigned to the rl_startup_hook variable (see section 2.3 Readline Variables).

Enums

BindType
Use for calls to generic_bind.

Functions

bind_key
Binds key to f in the currently active keymap. Returns non-zero in the case of an invalid key.

bind_key_if_unbound
Binds key to function f if it is not already bound in the currently active keymap. Returns non-zero in the case of an invalid key or if key is already bound.

bind_key_if_unbound_in_map
Binds key to function f if it is not already bound in map. Returns non-zero in the case of an invalid key or if key is already bound.

bind_key_in_map
Bind key to function f in map. Returns non-zero in the case of an invalid key.

bind_keyseq
Bind the key sequence represented by the string keyseq to the function f, beginning in the current keymap. This makes new keymaps as necessary. The return value is non-zero if keyseq is invalid.

bind_keyseq_if_unbound
Binds keyseq to function f if it is not already bound in the currently active keymap. Returns non-zero in the case of an invalid keyseq or if keyseq is already bound.

bind_keyseq_if_unbound_in_map
Binds keyseq to function f if it is not already bound in map. Returns non-zero in the case of an invalid keyseq or if keyseq is already bound.

bind_keyseq_in_map
Bind the key sequence represented by the string keyseq to the function f. This makes new keymaps as necessary. Initial bindings are performed in map. The return value is non-zero if keyseq is invalid.

generic_bind
Bind the key sequence represented by the string keyseq to the function f. type says what kind of data is pointed to by data; this can be a function (IsFunc), a macro (IsMacr), or a keymap (IsKmap). This makes new keymaps as necessary. The initial keymap in which to do bindings is map.

parse_and_bind
Parse line as if it had been read from the inputrc file and perform any key bindings and variable assignments found (see section 1.3 Readline Init File).

read_init_file
Read keybindings and variable assignments from path (see section 1.3 Readline Init File).

set_key
Equivalent to rl_bind_keyseq_in_map.

unbind_command_in_map
Unbind all keys that are bound to cmd in map.

unbind_function_in_map
Unbind all keys that execute function in map.

unbind_key
Bind key to the null function in the currently active keymap. Returns non-zero in case of error. Binds key to function f if it is not already bound in map. Returns non-zero in the case of an invalid key or if key is already bound.

unbind_key_in_map
Bind key to the null function in map. Returns non-zero in case of error.

Type Definitions

BindResult
Result type returned by all binding functions.