Function rl_sys::readline::binding::bind_keyseq_if_unbound

pub fn bind_keyseq_if_unbound(keyseq: &str, f: *mut Option<CommandFunction>) -> BindResult

Binds keyseq to function f if it is not already bound in the currently active keymap. Returns non-zero in the case of an invalid keyseq or if keyseq is already bound.

Examples

use libc::c_int;
use rl_sys::readline::{binding, util};

util::init();

extern "C" fn test_cmd_func(_count: c_int, _key: c_int) -> c_int {
  0
}

match binding::bind_keyseq_if_unbound("C-z", &mut Some(test_cmd_func)) {
    Ok(res) => assert!(res == 0),
    Err(_)  => assert!(false),
}

match binding::bind_keyseq_if_unbound("C-z", &mut Some(test_cmd_func)) {
    Ok(_)  => assert!(false),
    Err(_) => assert!(true),
}