Function rl_sys::readline::binding::parse_and_bind

pub fn parse_and_bind(line: &str) -> BindResult

Parse line as if it had been read from the inputrc file and perform any key bindings and variable assignments found (see section 1.3 Readline Init File).

Examples

use rl_sys::readline::{binding, util};

util::init();

match binding::parse_and_bind("\"\\ew\":\"\\C-e # macro\"") {
    Ok(res) => assert!(res == 0),
    Err(_)  => assert!(false),
}