Function rl_sys::readline::callback_handler_install

pub fn callback_handler_install(p: &str, lhandler: *mut Option<HandlerFunction>) -> Result<(), ReadlineError>

2.4.12 Alternate Interface

An alternate interface is available to plain readline()`. Some applications need to interleave keyboard I/O with file, device, or window system I/O, typically by using a main loop to select() on various file descriptors. To accommodate this need, readline can also be invoked as a 'callback' function from an event loop. There are functions available to make this easy.

Set up the terminal for readline I/O and display the initial expanded value of prompt p. Save the value of lhandler to use as a handler function to call when a complete line of input has been entered. The handler function receives the text of the line as an argument.