Module rl_sys::readline::funmap

2.4.4 Associating Function Names and Bindings

These functions allow you to find out what keys invoke named functions and the functions invoked by a particular key sequence. You may also associate a new function name with an arbitrary function.

Functions

add_funmap_entry
Add name to the list of bindable Readline command names, and make function f the function to be called when name is invoked.

function_dumper
Print the readline function names and the key sequences currently bound to them to rl_outstream. If readable is true, the list is formatted in such a way that it can be made part of an inputrc file and re-read.

function_of_keyseq
Return the function invoked by keyseq in keymap map. If map is None, the current keymap is used. If add_type is true, the type of the object is returned (one of IsFunc, IsKmap, or IsMacr).

invoking_keyseqs
Return an array of strings representing the key sequences used to invoke function in the current keymap.

invoking_keyseqs_in_map
Return an array of strings representing the key sequences used to invoke function in the current keymap.

list_funmap_names
Print the names of all bindable Readline functions to rl_outstream.

named_function
Return the function with name name.