Function rl_sys::readline::funmap::invoking_keyseqs_in_map

pub fn invoking_keyseqs_in_map(f: *mut Option<CommandFunction>, map: Keymap) -> Result<Vec<String>, ReadlineError>

Return an array of strings representing the key sequences used to invoke function in the current keymap.

Examples

use rl_sys::readline::{funmap, keymap, util};

util::init();

let cmd = funmap::named_function("self-insert").unwrap();
let km = keymap::get().unwrap();
let names = funmap::invoking_keyseqs_in_map(cmd, km).unwrap();
assert!(names.len() > 0);