Module rl_sys::readline::keymap

2.4.2 Selecting A Keymap

Key bindings take place on a keymap. The keymap is the association between the keys that the user types and the functions that get run. You can make your own keymaps, copy existing keymaps, and tell Readline which keymap to use.

Functions

copy
Return a new keymap which is a copy of map.

create_empty
Returns a new, empty keymap. The space for the keymap is allocated with malloc(); the caller should free it by calling rl_free_keymap() when done.

discard
Free the storage associated with the data in keymap. The caller should free keymap.

free
Free all storage associated with keymap. This calls rl_discard_keymap to free subordindate keymaps and macros.

get
Returns the currently active keymap.

get_by_name
Return the keymap matching name. name is one which would be supplied in a set keymap inputrc line (see section 1.3 Readline Init File).

get_name
Return the name matching map. name is one which would be supplied in a set keymap inputrc line (see section 1.3 Readline Init File).

make
Return a new keymap with the printing characters bound to rl_insert, the lowercase Meta characters bound to run their equivalents, and the Meta digits bound to produce numeric arguments.

set
Makes keymap the currently active keymap.