Function rl_sys::readline::misc::macro_dumper

pub fn macro_dumper(readable: bool)

Print the key sequences bound to macros and their values, using the current keymap, to rl_outstream. If readable is true, the list is formatted in such a way that it can be made part of an inputrc file and re-read.

Examples

use rl_sys::readline::{misc, util};

util::init();

misc::macro_dumper(true);