Module rl_sys::readline::modtext

2.4.7 Modifying Text

Functions

copy_text
Return a copy of the text between start and end in the current line.

delete_text
Delete the text between start and end in the current line. Returns the number of characters deleted.

insert_text
Insert text into the line at the current cursor position. Returns the number of characters inserted.

kill_text
Copy the text between start and end in the current line to the kill ring, appending or prepending to the last kill if the last command was a kill command. The text is deleted. If start is less than end, the text is appended, otherwise prepended. If the last command was not a kill, a new kill ring slot is used.

push_macro_input
Cause macro m to be inserted into the line, as if it had been invoked by a key bound to a macro. Not especially useful; use rl_insert_text() instead.