Function rl_sys::readline::modtext::kill_text

pub fn kill_text(start: i32, end: i32) -> Result<i32, ReadlineError>

Copy the text between start and end in the current line to the kill ring, appending or prepending to the last kill if the last command was a kill command. The text is deleted. If start is less than end, the text is appended, otherwise prepended. If the last command was not a kill, a new kill ring slot is used.

Examples

use rl_sys::readline::{modtext, util};

util::init();

assert!(modtext::insert_text("inserted").is_ok());
assert!(modtext::kill_text(5, 2).is_ok());