Module rl_sys::readline::naming

2.4.1 Naming a Function

The user can dynamically change the bindings of keys while using Readline. This is done by representing the function with a descriptive name. The user is able to type the descriptive name when referring to the function. Thus, in an init file, one might find

Meta-Rubout: backward-kill-word

This binds the keystroke Meta-Rubout to the function descriptively named backward-kill-word. You, as the programmer, should bind the functions you write to descriptive names as well.

Functions

add_func
Add name to the list of named functions. Make f be the function that gets called. If key is not -1, then bind it to function using rl_bind_key().