Module rl_sys::readline::undo

2.4.5 Allowing Undoing

Supporting the undo command is a painless thing, and makes your functions much more useful. It is certainly easy to try something if you know you can undo it.

If your function simply inserts text once, or deletes text once, and uses rl_insert_text() or rl_delete_text() to do it, then undoing is already done for you automatically.

If you do multiple insertions or multiple deletions, or any combination of these operations, you should group them together into one operation. This is done with rl_begin_undo_group() and rl_end_undo_group().

The types of events that can be undone are:

enum undo_code { UNDO_DELETE, UNDO_INSERT, UNDO_BEGIN, UNDO_END };

Notice that UNDO_DELETE means to insert some text, and UNDO_INSERT means to delete some text. That is, the undo code tells what to undo, not how to undo it. UNDO_BEGIN and UNDO_END are tags added by rl_begin_undo_group() and rl_end_undo_group().

Enums

UndoType
Undo Event Types

Functions

add_undo
Remember how to undo an event (according to what).

begin_undo_group
Begins saving undo information in a group construct. The undo information usually comes from calls to rl_insert_text() and rl_delete_text(), but could be the result of calls to rl_add_undo().

do_undo
Undo the first thing on the undo list. Returns an Err if there was nothing to undo, Ok if something was undone. Note that this depends on history being initialized and used.

end_undo_group
Closes the current undo group started with rl_begin_undo_group(). There should be one call to rl_end_undo_group() for each call to rl_begin_undo_group().

free_undo_list
Free the existing undo list. Note that this depends on history being initialized and used.

modifying
Tell Readline to save the text between start and end as a single undo unit. It is assumed that you will subsequently modify that text.