Function rl_sys::readline::undo::modifying

pub fn modifying(start: i32, end: i32) -> i32

Tell Readline to save the text between start and end as a single undo unit. It is assumed that you will subsequently modify that text.

If you neither insert nor delete text, but directly modify the existing text (e.g., change its case), call rl_modifying() once, just before you modify the text. You must supply the indices of the text range that you are going to modify.

Examples

use rl_sys::readline::{undo, util, vars};
use std::ffi::CString;

util::init();

let buffer = CString::new("test").unwrap().into_raw();
unsafe {
    vars::rl_line_buffer = buffer;
    let _ = CString::from_raw(buffer);
    assert!(undo::modifying(0, 1) == 0);
    let new_buffer = CString::new("Test").unwrap().into_raw();
    vars::rl_line_buffer = new_buffer;
    let _ = CString::from_raw(new_buffer);
}