Function rl_sys::readline::util::restore_state

pub fn restore_state(state: &mut ReadlineState) -> Result<i32, ReadlineError>

Restore Readline's internal state to that stored in state, which must have been saved by a call to save_state. The contents of the readline_state structure are documented in 'readline.h'. The caller is responsible for freeing the structure.

Examples

use rl_sys::readline::{ReadlineState, redisplay, util};
use std::default::Default;

util::init();

let mut state: ReadlineState = Default::default();
assert!(util::save_state(&mut state).is_ok());
assert!(state.prompt.is_null());
assert!(redisplay::set_prompt("promptly: ").is_ok());
let mut state2: ReadlineState = Default::default();
assert!(util::save_state(&mut state2).is_ok());
assert!(!state2.prompt.is_null());
assert!(util::restore_state(&mut state).is_ok());
assert!(state.prompt.is_null());